refactor: data request mappers
Refactor the DMS DTO mapper and their usage to allow a more simplified
way to add additional data types.

Create a parameterised enum which can be selected by TCS DTO class and
provides access to the appropriate schema, table and mapper to convert
it in to a DMS DTO.
Refactor DmsRecordAssembler to make use of the new enum to generically
retrieve and use the correct mapper class.
Create DmsMapper and refactor existing mappers to provide the ability to
call any mapper using the same method call.

TIS21-2997
TIS21-2999