refactor: address data request service comments
Remove the static import for `singletonList()` and `emptyList()`.
Move `Person` to be the first record added to the full sync list, as all
sent records use the same partition key this should mean that `Person`
is received first.
Document the the exclusion of `PersonOwner` from the full sync.

TIS21-2997
TIS21-2999