feat: Log and Slack on an exception ending the Job
The style of catch-log-throw is flagged as a smell.
Moving the event publishing to catch exceptions anywhere in the job, not
 just within the context of a chunk.
AFAIK, an exception in the chunk is rethrown and ends the job so that
 behavior is preserved.

fix: Use simple error messages passed to the Client

The message in the IllegalArgumentException is converted to JSON.
It is then read by the web client and needs to be a simple string.

chore: docs and simplifying RecordResendingJob

TIS21-3311: PersonStatus not updated when current programme is deleted