fix: treat requested PM amended date as instant
The trainee sync service expects the ProgrammeMembership amendedDate
field to be a timestamp, similar to the raw database data received via
DMS.

When fulfilling a request for a PM, the LocalDateTime is used from the
TCS DTO and is directly converted to a string.
This results in a string format with no zone information, causing
failures in trainee sync.

Update the ProgrammeMembershipMapper to handle LocalDateTime to Instant
conversion (assuming UTC, since there is no other information
available).
Update ProgrammeMembershipDmsDto to use appropriate date types for
`programmeStartDate`, `programmeEndDate` and `amendedDate`.

TIS21-4173
TIS21-2399