poc: include TIS metadata in data request messages
Concept is to add this metadata to allow tis-trainee-sync to determine
if the event is due to e.g. a rejected update, so that it can pass that
information on to the notifications service as well as performing the usual data
processing.

tis-tcs would generate the message when the GMC update fails and put it into the queue for
tis-sync. tis-sync would create the 'load' message with the TIS metadata
and push it into kinesis where it will be driven to the tis-trainee-sync
Record queue.

TIS21-6639